bridge rev lookup
