+++ execution_engine_testing/tests/src/test/explorer/faucet.rs
@@ -663,7 +663,7 @@ fn faucet_costs() {
     // This test will fail if execution costs vary.  The expected costs should not be updated
     // without understanding why the cost has changed.  If the costs do change, it should be
     // reflected in the "Costs by Entry Point" section of the faucet crate's README.md.
-    const EXPECTED_FAUCET_INSTALL_COST: u64 = 146_883_061_016;
+    const EXPECTED_FAUCET_INSTALL_COST: u64 = 158_323_799_135;
 
     const EXPECTED_FAUCET_SET_VARIABLES_COST: u64 = 79_640_385;
